pyproject.toml
[tool.poetry]
name = "polus_hack_back-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = {extras = ["mypy"], version = "^1.4.39"}
fastapi = "^0.79.0"
pydantic = {extras = ["email-validator", "email", "dotenv"], version = "^1.9.1"}
uvicorn = "^0.18.2"
psycopg2-binary = "^2.9.3"
SQLAlchemy-serializer = "^1.4.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
loguru = "^0.6.0"
pydantic-sqlalchemy = "^0.0.9"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
